Se realizo una implementación de llave secreta y seguridad de doble encriptacion

+++ Flask/listener/listener.py
@@ -1,8 +1,15 @@
 from app import app
 from flask import session,redirect,url_for,request,jsonify,render_template
 import model.model as model
-#Import Model
-#import models.indexModel as indexModel
+import os,secrets,base64,json
+
+#   DO: Return key and user from request
+#   @param 
+#   @return json user,key
+def uncryptedpKey(encryptedkey):
+    decodeKeyAndUser = base64.b64decode(request.headers.get('key')).decode("utf-8")
+    splittedCode = decodeKeyAndUser.split('USER:')
+    return {'user':splittedCode[1],'key':base64.b64decode(splittedCode[0]).decode("utf-8")}
 
 #   DO: render template with info
 #   @param 
@@ -17,8 +24,11 @@ def index():
 @app.route("/getuser",methods = ['GET'])
 def getuser():
     try:
-        response = model.get_users()
-        return jsonify({'data':response})
+        jsonKey = uncryptedpKey(request.headers.get('key'))
+        if(jsonKey['key'] == app.secret_key):
+            response = model.get_users()
+            return jsonify({'data':response})
+        return jsonify({'error':'not is the same key'})
     except:
         return jsonify({'error':'we have some error'})
 
@@ -28,10 +38,12 @@ def getuser():
 @app.route("/postuser",methods = ['POST'])
 def postuser():
     try:
-        print(request.json)
-        idUser = request.json.get('id')
-        response = model.get_one_user(idUser)
-        return jsonify({'data':response})
+        jsonKey = uncryptedpKey(request.headers.get('key'))
+        if(jsonKey['key'] == app.secret_key):
+            idUser = request.json.get('id')
+            response = model.get_one_user(idUser)
+            return jsonify({'data':response})
+        return jsonify({'error':'not is the same key'})
     except:
         return jsonify({'error':'we have some error'})
 
@@ -41,11 +53,14 @@ def postuser():
 @app.route("/putuser", methods=['PUT'])
 def putuser():
     try:
-        name = request.json.get('name')
-        birthdate = request.json.get('birthdate')
-        job = request.json.get('job')
-        response = model.insert_user(name,birthdate,job)
-        return jsonify({'data':response})
+        jsonKey = uncryptedpKey(request.headers.get('key'))
+        if(jsonKey['key'] == app.secret_key):
+            name = request.json.get('name')
+            birthdate = request.json.get('birthdate')
+            job = request.json.get('job')
+            response = model.insert_user(name,birthdate,job)
+            return jsonify({'data':response})
+        return jsonify({'error':'not is the same key'})
     except:
         return jsonify({'error':'we have some error'})
 
@@ -55,8 +70,29 @@ def putuser():
 @app.route("/deleteuser", methods=['DELETE'])
 def deleteuser():
     try:
-        idUser = request.json.get('id')
-        response = model.delete_user(idUser)
-        return jsonify({'data':response})
+        jsonKey = uncryptedpKey(request.headers.get('key'))
+        if(jsonKey['key'] == app.secret_key):
+            idUser = request.json.get('id')
+            response = model.delete_user(idUser)
+            return jsonify({'data':response})
+        return jsonify({'error':'not is the same key'})
     except:
         return jsonify({'error':'we have some error'})
+
+#   DO: Return secretkey
+#   @param
+#   @return json data
+@app.route("/getsecretkey", methods=['POST'])
+def getsecretkey():
+    #session['user'] = 'enrique'
+    #session['pws'] = 'flask'
+    #session['key'] = base64.b64encode(app.secret_key.encode('ascii')).decode("utf-8")
+    return jsonify({'key':base64.b64encode(app.secret_key.encode('ascii')).decode("utf-8")})
+    
+#   DO: Return responsecode
+#   @param
+#   @return json data
+@app.route("/getresponsekey", methods=['POST'])
+def getresponsekey():
+    key=request.json.get('key')+'USER:'+request.json.get('user')
+    return jsonify({'key':base64.b64encode(key.encode('ascii')).decode("utf-8")})
